The new interface, markup and styles have been transferred to BEM.
The markup of HTML layouts and site pages has been changed for the BEM methodology.
Updated CSS for new HTML classes.
Header and footer styles are now in separate files.
The navigation menu is now on the left instead of under the header.
New icon for users without an avatar.

<!> Dark theme does not work.